add native-debian-own-riscv.ini
WARNING: not usable as is, just for reference.
It contains hard coded paths. It is just to demostrate how to add a
RISC-V toolchain to compile IronOS.

Add the RISC-V installation with (with adapted paths):

git clone https://github.com/riscv/riscv-gnu-toolchain
cd riscv-gnu-toolchain
./configure --prefix=/srv/scratch/entrup/riscv-toolchain-install --enable-multilib
make -j$(nproc)

TODO: the right place for that would be the toolchains subproject